requirements.txt
langchain-text-splitters
langchain-community
langgraph
python-dotenv
beautifulsoup4
langchain-google-genai